synced mana bar properly to fix #12
presumably health bar too but i never really noticed it there

+++ src/main/java/io/github/moosyu/util/UnshatteredUtils.java
@@ -304,7 +304,7 @@ public final class UnshatteredUtils {
     public static boolean passesManaCheck(Player player, int manaCost) {
         double playerManaAmount = player.getData(UnshatteredAttachments.PLAYER_STATE.get()).getCurrentStat(PlayerStateAttachment.Stat.MANA);
         if (playerManaAmount < manaCost) {
-            player.sendSystemMessage(Component.literal("You don't have enough mana to use this " + "(" + playerManaAmount + "/" + manaCost + ").").withColor(ERROR_COLOR));
+            player.sendSystemMessage(Component.literal("You don't have enough mana to use this " + "(" + (int) playerManaAmount + "/" + manaCost + ").").withColor(ERROR_COLOR));
             return false;
         }
         return true;
